FAB-14868 Update pipeline scripts
This patch updates the existing pipeline scripts to use
global shared library and removes the redundant code from
this repo.

Attempt to not pull images we do not use directly, but that
integration framework depends upon existing. Pull a tiny busybox image
instead and tag it as those images if they do not already exist.

Change-Id: I1098a4828143ec6b677867bd863e26123e274877
Signed-off-by: rameshthoomu <rameshbabu.thoomu@gmail.com>
Signed-off-by: Morgan Bauer <mbauer@us.ibm.com>