FAB-16697 remove linters
go lint is not run. Remove go lint.

go fmt and goimports do the same thing. goimports is more powerful due
to import reordering, so rely only on goimports. Remove go fmt.

The build cache is not the build cache we use for building a binary,
so we end up building everything multiple times instead of using the
cache. Remove the cache definition.

go vet is intelligent and ignores the vendor directory on it's
own. Take it out of the loop and invoke it once at the top level.

Change-Id: Ieb057298f6ecd83b67607de29c24572a4c35018a
Signed-off-by: Morgan Bauer <mbauer@us.ibm.com>